refactor: 💡 Renamve validator.ts to envalid.ts
