+++ src/cmd/bootstrap.ts
@@ -4,9 +4,9 @@ import { copyFile } from 'fs/promises'
 import { Argv } from 'yargs'
 import { $ } from 'zx'
 import { OtomiDebugger, terminal } from '../common/debug'
+import { env } from '../common/envalid'
 import { BasicArguments, loadYaml, setParsedArgs } from '../common/no-deps'
 import { cleanupHandler, otomi } from '../common/setup'
-import { env } from '../common/validators'
 import { genSops } from './gen-sops'
 
 const fileName = 'bootstrap'
