+++ src/cmd/check-policies.ts
@@ -2,10 +2,10 @@ import { rmSync } from 'fs'
 import { Argv } from 'yargs'
 import { $, nothrow } from 'zx'
 import { OtomiDebugger, terminal } from '../common/debug'
+import { env } from '../common/envalid'
 import { hfTemplate } from '../common/hf'
 import { loadYaml, LOG_LEVEL, LOG_LEVELS, setParsedArgs } from '../common/no-deps'
 import { cleanupHandler, otomi, PrepareEnvironmentOptions } from '../common/setup'
-import { env } from '../common/validators'
 import { Arguments, helmOptions } from '../common/yargs-opts'
 
 const fileName = 'check-policies'
