+++ src/cmd/pull.ts
@@ -1,9 +1,9 @@
 import { Argv } from 'yargs'
 import { $ } from 'zx'
 import { OtomiDebugger, terminal } from '../common/debug'
+import { env } from '../common/envalid'
 import { setParsedArgs } from '../common/no-deps'
 import { cleanupHandler, otomi, PrepareEnvironmentOptions } from '../common/setup'
-import { env } from '../common/validators'
 import { Arguments as HelmArgs, helmOptions } from '../common/yargs-opts'
 import { Arguments as BootsrapArgs, bootstrap } from './bootstrap'
