+++ src/common/gitea-push.ts
@@ -1,9 +1,9 @@
 import { existsSync } from 'fs'
 import { $, cd, nothrow } from 'zx'
 import { terminal } from './debug'
+import { env } from './envalid'
 import { hfValues } from './hf'
 import { waitTillAvailable } from './no-deps'
-import { env } from './validators'
 
 export const giteaPush = async (): Promise<void> => {
   const debug = terminal('Gitea Push')
