+++ src/common/hf.ts
@@ -1,8 +1,8 @@
 import { dump, load } from 'js-yaml'
 import { Transform } from 'stream'
 import { $, ProcessOutput, ProcessPromise } from 'zx'
+import { env } from './envalid'
 import { asArray, LOG_LEVELS } from './no-deps'
-import { env } from './validators'
 import { Arguments } from './yargs-opts'
 import { ProcessOutputTrimmed, Streams } from './zx-enhance'
